library/requirements.txt
django
django-allauth
django-debug-toolbar
django-environ
django-extensions

graphene-django

pillow
psycopg2-binary

flake8
isort